created new MapPage
